[PaTH] Use int64 addressing across remaining kernels for long sequences
Follow-up to #769: the remaining parallel/intra/preprocess/transform kernels
still used int32 bos/eos offsets that overflow at large T*HQ*D. Cast bos/eos to
int64 at assignment, keeping sequence lengths int32 for block-pointer shapes.

+++ fla/ops/path_attn/intra_chunk_preprocess_bwd_prepare.py
@@ -57,11 +57,11 @@ def chunk_transform_qk_bwd_kernel_prepare(
 
     if IS_VARLEN:
         i_n, i_t = tl.load(indices + i_t * 2).to(tl.int32), tl.load(indices + i_t * 2 + 1).to(tl.int32)
-        bos, eos = tl.load(offsets + i_n).to(tl.int32), tl.load(offsets + i_n + 1).to(tl.int32)
-        T = eos - bos
+        bos, eos = tl.load(offsets + i_n).to(tl.int64), tl.load(offsets + i_n + 1).to(tl.int64)
+        T = (eos - bos).to(tl.int32)
         boh = tl.load(chunk_offsets + i_n).to(tl.int32)
     else:
-        bos, eos = i_n * T, i_n * T + T
+        bos, eos = (i_n * T).to(tl.int64), (i_n * T + T).to(tl.int64)
         NT = tl.cdiv(T, BT)
         boh = i_n * NT
